Рефакторинг: выносит список комментариев и блок добавления нового комментария в отдельные смарт-компоненты
