--- src/components/comment.js
@@ -1,34 +1,0 @@
-import AbstractComponent from "./abstract-component.js";
-import moment from "moment";
-
-const createCommentTemplate = (comment) => {
-  const {author, comment: text, date, emotion} = comment;
-  const commentDate = moment(date).fromNow();
-  return (
-    `<li class="film-details__comment">
-        <span class="film-details__comment-emoji">
-          <img src="./images/emoji/${emotion}.png" width="55" height="55" alt="emoji-${emotion}">
-        </span>
-        <div>
-          <p class="film-details__comment-text">${text}</p>
-          <p class="film-details__comment-info">
-            <span class="film-details__comment-author">${author}</span>
-            <span class="film-details__comment-day">${commentDate}</span>
-            <button class="film-details__comment-delete">Delete</button>
-          </p>
-        </div>
-      </li>`
-  );
-};
-
-export default class Comment extends AbstractComponent {
-  constructor(comment) {
-    super();
-    this._comment = comment;
-  }
-
-  getTemplate() {
-    return createCommentTemplate(this._comment);
-  }
-}
-
