+++ src/components/film-popup.js
@@ -23,30 +23,12 @@ const createButtonMarkup = (name, isChecked) => {
   );
 };
 
-const createAddEmojiMarkup = (emotion) => {
-  return emotion ? `<img src="images/emoji/${emotion}.png" width="55" height="55" alt="emoji-${emotion}">` : ``;
-};
-
-const createEmojiListMarkup = () => {
-  return EMOJIS.map((emotion) => {
-    return (
-      `<input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-${emotion}" value="${emotion}">
-        <label class="film-details__emoji-label" for="emoji-${emotion}">
-          <img src="./images/emoji/${emotion}.png" width="30" height="30" alt="emoji">
-        </label>`
-    );
-  }).join(`\n`);
-};
-
 const getGenreTitle = (genres) => genres.length > 1 ? `Genres` : `Genre`;
-const getCommentsTitle = (commentsCount) => commentsCount > 1 ? `Comments` : `Comment`;
 
-const createFilmPopupTemplate = (film, emotion) => {
+const createFilmPopupTemplate = (film) => {
   const {title, alternativeTitle, totalRating, poster, ageRating, director, writers, actors, releaseCountry, genres, description} = film;
   const genresMarkup = createGenresMarkup(genres);
   const genresTitle = getGenreTitle(genres);
-  const commentsCount = film.comments.length;
-  const commentsTitle = getCommentsTitle(commentsCount);
   const addWatchListButton = createButtonMarkup(`watchlist`, film.watchlist);
   const watchedButton = createButtonMarkup(`watched`, film.alreadyWatched);
   const favoriteButton = createButtonMarkup(`favorite`, film.favorite);
@@ -123,28 +105,7 @@ const createFilmPopupTemplate = (film, emotion) => {
       </section>
     </div>
 
-    <div class="form-details__bottom-container">
-      <section class="film-details__comments-wrap">
-        <h3 class="film-details__comments-title">${commentsTitle} <span class="film-details__comments-count">${commentsCount}</span></h3>
 
-        <ul class="film-details__comments-list">
-        </ul>
-
-        <div class="film-details__new-comment">
-          <div for="add-emoji" class="film-details__add-emoji-label">
-            ${createAddEmojiMarkup(emotion)}
-           </div>
-
-          <label class="film-details__comment-label">
-            <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
-          </label>
-
-          <div class="film-details__emoji-list">
-            ${createEmojiListMarkup()}
-          </div>
-        </div>
-      </section>
-    </div>
   </form>
 </section>`
   );
@@ -158,8 +119,6 @@ export default class FilmPopup extends AbstractSmartComponent {
     this._addWatchListChangeHandler = null;
     this._WatchedChangeHandler = null;
     this._favoriteChangeHandler = null;
-    this._emotion = ``;
-    this._subscribeOnEvents();
   }
 
   getTemplate() {
@@ -171,7 +130,6 @@ export default class FilmPopup extends AbstractSmartComponent {
     this.setAddWatchListButtonChangeHandler(this._addWatchListChangeHandler);
     this.setWatchedButtonChangeHandler(this._WatchedChangeHandler);
     this.setFavoriteButtonChangeHandler(this._favoriteChangeHandler);
-    this._subscribeOnEvents();
   }
 
   rerender() {
@@ -196,18 +154,5 @@ export default class FilmPopup extends AbstractSmartComponent {
   setFavoriteButtonChangeHandler(handler) {
     this.getElement().querySelector(`#favorite`).addEventListener(`change`, handler);
     this._favoriteChangeHandler = handler;
-  }
-
-  reset() {
-    this._emotion = ``;
-    this.rerender();
-  }
-
-  _subscribeOnEvents() {
-    const element = this.getElement();
-    element.querySelector(`.film-details__emoji-list`).addEventListener(`change`, (evt) => {
-      this._emotion = evt.target.value;
-      this.rerender();
-    });
   }
 }
