+++ src/models/comments.js
@@ -1,8 +1,6 @@
 export default class Comments {
   constructor() {
     this._comments = [];
-
-    this._dataChangeHandlers = [];
   }
 
   getComments() {
@@ -11,7 +9,6 @@ export default class Comments {
 
   setComments(comments) {
     this._comments = Array.from(comments);
-    this._callHandlers(this._dataChangeHandlers);
   }
 
   getCommentsByIds(ids) {
@@ -26,21 +23,10 @@ export default class Comments {
     }
 
     this._comments = [].concat(this._comments.slice(0, index), this._comments.slice(index + 1));
-    this._callHandlers(this._dataChangeHandlers);
-
     return true;
   }
 
   addComment(comment) {
     this._comments = [].concat(comment, this._comments);
-    this._callHandlers(this._dataChangeHandlers);
-  }
-
-  setDataChangeHandler(handler) {
-    this._dataChangeHandlers.push(handler);
-  }
-
-  _callHandlers(handlers) {
-    handlers.forEach((handler) => handler());
   }
 }
